rename tree_hash.rb
