+++ 2017/DmLaziuk/1/README.md
@@ -51,7 +51,7 @@ $ NAME=mel ruby trees.rb
 После вывода каждого из деревьев автоматически определяет действие, которое нужно выполнить с данным деревом: обрезать (если глубина дерева больше 5), срубить  (если сумма узлов более 5000) или оставить.
 
 ### Рефакторинг
-Файлы: treehash.rb, trees4.rb, trees.zip
+Файлы: tree_hash.rb, trees4.rb, trees.zip
 
 Все методы для работы с деревом вынесены в отдельный класс TreeHash.
