+++ 2017/DmLaziuk/1/tree_hash.rb
